+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## 1.5.0 (Unreleased)
 ## 1.4.0 (April 12, 2019)
 
 ENHANCEMENTS: